tablet: specify the tool button numbers are simply sequentially numbered
We're using sequentially numbered buttons for the pad because actual tablets
are quite varied in how the present buttons (BTN_A, BTN_0, etc.). For this
reason, libinput numbers pad buttons sequentially.

Let's do the same for tablet tools. Unfortunately libinput doesn't do this, so
the mapping is up to the compositor but at least we can get the wayland
protocol right. The ony event codes a tool currently sends are BTN_STYLUS and
BTN_STYLUS2 anyway, so it's already sequentially numbered with a make-pretend
fancy name.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Yong Bakos <ybakos@humanoriented.com>